resolved all compiler warnings.

+++ Tunnelour/Tunnelour/include/Text_Component.h
@@ -45,24 +45,24 @@ class Text_Component: public Tunnelour::Bitmap_Component {
                             yoffset(-1),
                             xadvance(-1) {
     }
-    float id;
-    float x;
-    float y;
-    float width;
-    float height;
-    float xoffset;
-    float yoffset;
-    float xadvance;
+    double id;
+    double x;
+    double y;
+    double width;
+    double height;
+    double xoffset;
+    double yoffset;
+    double xadvance;
   };
 
   struct Font {
-    float image_width;
-    float image_height;
+    double image_width;
+    double image_height;
     // This is the distance in pixels between each line of text.
-    float line_height;
+    double line_height;
     std::string font_name;
     std::wstring font_texture_name;
-    float character_widths[256];
+    double character_widths[256];
     Raw_Character_Frame raw_character_frames[256];
     Vertex_Type character_frames[256][6];
     D3DXCOLOR font_color;
